+++ tests/testthat/test_mlr_task_generators_simsurv.R
@@ -1,6 +1,0 @@
-skip_if_not_installed("simsurv")
-test_that("mlr_task_generators_simsurv", {
-  gen = mlr_task_generators$get("simsurv")
-  task = gen$generate(100)
-  expect_task_surv(task)
-})