+++ NEWS.md
@@ -1,3 +1,12 @@
+# mlr3proba 0.6.3
+
+* Add new tasks from `survival` package: `veteran`, `pbc`, `mgus`, `gbsg`
+  * Refined docs for loaded tasks, `task$help()` works as it should now
+  * All loaded tasks have now **complete cases** (no features with missing values)
+* Refined docs for task generators
+* Added task generator from `coxed` package
+* Added new methods for `TaskSurv`: `cens_prop()`, `admin_cens_prop()`, `dep_cens_prop()` and `prop_haz()`
+
 # mlr3proba 0.6.2
 
 * Updates in `surv.cindex` measure
@@ -12,7 +21,7 @@
 # mlr3proba 0.6.0
 
 * Optimized `surv.logloss` and `calib_alpha` measures (bypassing `distr6`)
-* Update/refine all measure docs (namign conventions from upcoming scoring rules paper) + doc templates
+* Update/refine all measure docs (naming conventions from upcoming scoring rules paper) + doc templates
 * fix very rare bugs in `calib_alpha`, `surv.logloss` and `surv.graf` (version with proper = FALSE)
 
 # mlr3proba 0.5.9